[MOD] moved editor creation to a helper & seperated static methods from customeditorwindow for clarity
